Changed the Structure
My proposition:
-"public_html"/"htdocs": here are all the pages to be called.
-"includes" : here are all the files that do not exists as pages on
their own, but rather are included.
-"propose","proposed","voting","approved_profile" are all pages to be
viewed. The ".tpl"s are ONLY required/included. The pages are seen from
the ".php"s which run the includes on their ".tpl"s.
-Renamed the photo to "front end preview" from the gibrish-woobrish :D
-index.php/index.html will be redirecting to the front page (being for
now the voting.php). That is to have clear names (eg. not "index.php"
for the voting.).
------------------------------------------------
P.S.: Whatever had code in it I left it AS IT. Which means